AI/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
chromadb==0.4.24
google-generativeai==0.4.1
google-genai==0.3.0
openai==1.14.3
python-dotenv==1.0.1
requests==2.31.0
cryptography==42.0.5
pyarmor==8.5.5
